refactor: fromEntity() 등을 이용하여 DTO 생성하도록 수정
