+++ backend/was/src/chat/chat.service.ts
@@ -8,7 +8,7 @@ import { ERR_MSG } from 'src/common/constants/errors';
 import { LoggerService } from 'src/logger/logger.service';
 import { Member } from 'src/members/entities/member.entity';
 import { QueryFailedError, Repository } from 'typeorm';
-import { ChattingMessageResponseDto } from './dto/chatting-messag-response.dto';
+import { ChattingMessageResponseDto } from './dto/chatting-message-response.dto';
 import { ChattingRoomResponseDto } from './dto/chatting-room-response.dto';
 import { CreateChattingMessageDto } from './dto/create-chatting-message.dto';
 import { UpdateChattingRoomDto } from './dto/update-chatting-room.dto';
@@ -111,24 +111,17 @@ export class ChatService {
     const rooms: ChattingRoom[] = await this.chattingRoomRepository.findBy({
       id,
     });
-    return rooms.map((room: ChattingRoom) => {
-      const roomDto: ChattingRoomResponseDto = new ChattingRoomResponseDto();
-      roomDto.id = room.id;
-      roomDto.title = room.title;
-      return roomDto;
-    });
+    return rooms.map((room: ChattingRoom) =>
+      ChattingRoomResponseDto.fromEntity(room),
+    );
   }
 
   async findMessagesById(id: string): Promise<ChattingMessageResponseDto[]> {
     const messages: ChattingMessage[] =
       await this.chattingMessageRepository.findBy({ id });
-    return messages.map((message: ChattingMessage) => {
-      const messageDto = new ChattingMessageResponseDto();
-      messageDto.id = message.id;
-      messageDto.isHost = message.isHost;
-      messageDto.message = message.message;
-      return messageDto;
-    });
+    return messages.map((message: ChattingMessage) =>
+      ChattingMessageResponseDto.fromEntity(message),
+    );
   }
 
   async updateRoom(
